backend/requirements.txt
dj-rest-auth==2.2.5
django-allauth==0.51.0
django-crispy-forms==1.14.0
asgiref==3.5.2
aws==0.2.5
bcrypt==3.2.2
binary==1.0.0
blessed==1.19.1
boto==2.49.0
boto3==1.24.16
botocore==1.27.16
cement==2.8.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.3
distlib==0.3.4
Django==4.0.3
django-ses==3.1.0
django-smtp-ssl==1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
fabric==2.7.0
flake8==4.0.1
future==0.16.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
install==1.3.5
invoke==1.7.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
paramiko==2.11.0
pathlib2==2.3.7.post1
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.2
prettytable==3.3.0
psycopg==3.0.15
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==5.4.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.1.9
websocket-client==0.59.0
django-leaderboard==0.1.0
django-fullurl==1.3

